REFACTOR Ajuste nos namespaces
