+++ renda-fixa-infra/Repositories/ClienteRepository.cs
@@ -2,11 +2,11 @@
 using RendaFixa.Domain.Interfaces;
 using RendaFixa.Infrastruct.Context;
 
-namespace RendaFixa.Infrastruct.Repository;
+namespace RendaFixa.Infrastruct.Repositories;
 
 public class ClienteRepository : BaseRepository<Cliente>, IClienteRepository
 {
     public ClienteRepository(AppDbContext contexto) : base(contexto)
-    {}
+    { }
 
 }
