+++ renda-fixa-infra/Repositories/ContaRepository.cs
@@ -2,11 +2,11 @@
 using RendaFixa.Domain.Interfaces;
 using RendaFixa.Infrastruct.Context;
 
-namespace RendaFixa.Infrastruct.Repository;
+namespace RendaFixa.Infrastruct.Repositories;
 
 public class ContaRepository : BaseRepository<Conta>, IContaRepository
 {
     public ContaRepository(AppDbContext contexto) : base(contexto)
-    {}
+    { }
 
 }
