+++ renda-fixa-infra/Repositories/ProdutoRendaRepository.cs
@@ -4,7 +4,7 @@ using RendaFixa.Domain.Entities;
 using RendaFixa.Domain.Interfaces;
 using RendaFixa.Infrastruct.Context;
 
-namespace RendaFixa.Infrastruct.Repository;
+namespace RendaFixa.Infrastruct.Repositories;
 
 public class ProdutoRendaFixaRepository : BaseRepository<ProdutoRendaFixa>, IProdutoRendaFixaRepository
 {
