+++ renda-fixa-service/Config/ServiceExtensions.cs
@@ -2,13 +2,13 @@
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 
-namespace RendaFixa.Aplication.Config;
+namespace RendaFixa.Service.Config;
 public static class SeviceExtension
 {
-   public static void ConfigureApplicationApp(this IServiceCollection services)
-   {
+    public static void ConfigureApplicationApp(this IServiceCollection services)
+    {
         services.AddAutoMapper(Assembly.GetExecutingAssembly());
         services.AddMediatR(x => x.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly()));
         services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-   }
+    }
 }
