+++ renda-fixa-service/UseCases/Queries/GetClienteById/GetClienteByIdHandler.cs
@@ -18,7 +18,6 @@ public class GetClienteByIdHandler : IRequestHandler<GetClienteByIdRequest, GetC
 
     public async Task<GetClienteByIdResponse> Handle(GetClienteByIdRequest request, CancellationToken cancellationToken)
     {
-        var aporte = mapper.Map<Conta>(request);
 
         var resposta = await repository.GetByIdAsync(request.ClienteId, cancellationToken);
