+++ renda-fixa-service/UseCases/Queries/GetClienteById/GetClienteByIdMapper.cs
@@ -7,7 +7,7 @@ public sealed class GetClienteByIdMapper : Profile
 {
     public GetClienteByIdMapper()
     {
-        CreateMap<GetClienteByIdRequest, Cliente>();
+        //CreateMap<GetClienteByIdRequest, Cliente>();
         CreateMap<Cliente, GetClienteByIdResponse>();
     }
 }
