+++ renda-fixa-service/UseCases/Queries/GetClienteById/GetClienteByIdResponse.cs
@@ -2,5 +2,5 @@
 
 public sealed record GetClienteByIdResponse(
     int Id,
-    int Nome);
+    string Nome);
