+++ renda-fixa-service/UseCases/Queries/GetContaById/GetContaByIdHandler.cs
@@ -18,7 +18,6 @@ public class GetContaByIdHandler : IRequestHandler<GetContaByIdRequest, GetConta
 
     public async Task<GetContaByIdResponse> Handle(GetContaByIdRequest request, CancellationToken cancellationToken)
     {
-        var aporte = mapper.Map<Conta>(request);
 
         var resposta = await repository.GetByIdAsync(request.ContaId, cancellationToken);
